Updated Numbers Round Solver Data Structure
The Tree List in InclusionMap has been moved to the NumbersRoundSolver
as a deque object, used as a queue for the solve function.

The solve function was converted to a generator function, yielding
solutions. A get_all_solutions functions was created as a short-hand,
but uses the solve function for its functionality.